requirements.txt
setuptools >= 69.0.3
discord.py >= 2.3.2
python-dotenv >= 1.0.0